+++ plone/app/upgrade/v50/betas.py
@@ -3,6 +3,7 @@ from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.interfaces import IMailSchema
 from Products.CMFPlone.interfaces import IMarkupSchema
 from Products.CMFPlone.interfaces import ISecuritySchema
+from Products.CMFPlone.interfaces import IUserGroupsSettingsSchema
 from Products.CMFPlone.interfaces import ILanguageSchema
 from plone.app.linkintegrity.upgrades import migrate_linkintegrity_relations
 from plone.app.upgrade.utils import loadMigrationProfile
@@ -279,6 +280,81 @@ def upgrade_querystring(context):
     context.upgradeProfile('plone.app.querystring:default')
 
 
+def upgrade_usergroups_controlpanel_settings(context):
+    """Copy users and groups control panel settings from portal properties
+    into the new registry.
+    """
+
+    # get the old site properties
+    portal_url = getToolByName(context, 'portal_url')
+    portal = portal_url.getPortalObject()
+    portal_properties = getToolByName(portal, "portal_properties")
+    site_properties = portal_properties.site_properties
+
+    # get the new registry
+    registry = getUtility(IRegistry)
+
+    # XXX: Somehow this code is executed for old migration steps as well
+    # ( < Plone 4 ) and breaks because there is no registry. Looking up the
+    # registry interfaces with 'check=False' will not work, because it will
+    # return a settings object and then fail when we try to access the
+    # attributes.
+    try:
+        settings = registry.forInterface(IUserGroupsSettingsSchema,
+                                         prefix='plone')
+    except KeyError:
+        settings = False
+    if settings:
+        settings.many_groups = site_properties.getProperty('many_groups',
+                                                           False)
+        settings.many_users = site_properties.getProperty('many_users',
+                                                          False)
+
+
 def to50rc1(context):
     """5.0beta4 -> 5.0rc1"""
     loadMigrationProfile(context, 'profile-plone.app.upgrade.v50:to50rc1')
+
+    upgrade_usergroups_controlpanel_settings(context)
+
+    portal = getSite()
+    # Migrate settings from portal_properties to the configuration registry
+    pprop = getToolByName(portal, 'portal_properties')
+    site_properties = pprop['site_properties']
+
+    # These have been migrated in previous upgrade steps. Safe to remove.
+    properties_to_remove = ['allowAnonymousViewAbout',
+                            'available_editors',
+                            'default_contenttype',
+                            'default_editor',
+                            'disable_folder_sections',
+                            'disable_nonfolderish_sections',
+                            'enable_inline_editing',
+                            'enable_link_integrity_checks',
+                            'enable_livesearch',
+                            'enable_sitemap',
+                            'exposeDCMetaTags',
+                            'ext_editor',
+                            'forbidden_contenttypes',
+                            'lock_on_ttw_edit',
+                            'many_groups',
+                            'many_users',
+                            'number_of_days_to_keep',
+                            'search_collapse_options',
+                            'search_enable_batch_size',
+                            'search_enable_description_search',
+                            'search_review_state_for_anon',
+                            'search_enable_sort_on',
+                            'search_enable_title_search',
+                            'types_not_searched',
+                            'use_email_as_login',
+                            'use_folder_contents',
+                            'use_uuid_as_userid',
+                            'user_registration_fields',
+                            'visible_ids',
+                            'webstats_js']
+    for p in properties_to_remove:
+        if site_properties.hasProperty(p):
+            site_properties._delProperty(p)
+
+    import pdb; pdb.set_trace( )